+++ NEWS.md
@@ -2,6 +2,7 @@
 
 ## New features
 
+* Add ``g3_parameterized('x', by_step = TRUE)`` for seasonal parameters #115
 * Support R factor levels as well as MFDB-style attributes for likelihood observation data.
   This means dplyr ``group_by(length = cut(...))`` can be used as well as MFDB for groupings #112